fix(catalog): min_stock cast float antes de round — evita 500 en PHP 8.2

+++ app/Http/Controllers/CatalogController.php
@@ -107,7 +107,7 @@ class CatalogController extends Controller
         }
 
         if ($validated['sale_mode'] === 'unit') {
-            $validated['min_stock'] = (int) round($validated['min_stock'] ?? 0);
+            $validated['min_stock'] = (int) round((float) ($validated['min_stock'] ?? 0));
         }
 
         $product = Product::create([
@@ -172,7 +172,7 @@ class CatalogController extends Controller
         ]);
 
         if ($validated['sale_mode'] === 'unit') {
-            $validated['min_stock'] = (int) round($validated['min_stock'] ?? 0);
+            $validated['min_stock'] = (int) round((float) ($validated['min_stock'] ?? 0));
         }
 
         $updates = [
